Rename typings -> types
