+++ src/cow-react/common/pure/CurrencyInputPanel/CurrencyPreview.tsx
@@ -5,7 +5,7 @@ import { formatSmartAmount } from 'utils/format'
 import { FiatValue } from 'components/CurrencyInputPanel/FiatValue'
 import { Trans } from '@lingui/macro'
 import { PriceImpact } from 'hooks/usePriceImpact'
-import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/typings'
+import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
 
 interface BuiltItProps {
   className: string
