+++ src/cow-react/common/pure/CurrencyInputPanel/types.ts
