+++ src/cow-react/modules/limitOrders/containers/LimitOrdersConfirmModal/index.tsx
@@ -5,7 +5,7 @@ import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 import { CloseIcon } from 'theme'
 import { useAtomValue } from 'jotai/utils'
 import { useHigherUSDValue } from 'hooks/useStablecoinPrice'
-import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/typings'
+import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
 import { LimitOrdersConfirm } from '../../pure/LimitOrdersConfirm'
 import { tradeFlow, TradeFlowContext } from '../../services/tradeFlow'
 import { limitRateAtom } from '../../state/limitRateAtom'
