+++ src/cow-react/modules/limitOrders/pure/LimitOrdersConfirm/index.tsx
@@ -3,7 +3,7 @@ import { Trans } from '@lingui/macro'
 import { ButtonSize } from 'theme'
 import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
 import { ButtonPrimary } from 'components/Button'
-import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/typings'
+import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
 import { CurrencyPreview } from '@cow/common/pure/CurrencyInputPanel'
 import { LimitOrdersDetails } from '../LimitOrdersDetails'
 import { TradeFlowContext } from '../../services/tradeFlow'
