+++ src/cow-react/modules/mainMenu/types.ts
