+++ src/cow-react/modules/swap/containers/NewSwapWidget/index.tsx
@@ -18,7 +18,7 @@ import { useShowRecipientControls } from '@cow/modules/swap/hooks/useShowRecipie
 import usePriceImpact from 'hooks/usePriceImpact'
 import { useTradePricesUpdate } from '@cow/modules/swap/hooks/useTradePricesUpdate'
 import { useCurrencyBalance } from 'state/connection/hooks'
-import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/typings'
+import { CurrencyInfo } from '@cow/common/pure/CurrencyInputPanel/types'
 import { Field } from 'state/swap/actions'
 import { tokenViewAmount } from '@cow/modules/swap/helpers/tokenViewAmount'
 import { useHigherUSDValue } from 'hooks/useStablecoinPrice'
@@ -27,7 +27,7 @@ import React, { useState } from 'react'
 import { useModalIsOpen } from 'state/application/hooks'
 import { ApplicationModal } from 'state/application/reducer'
 import { useSwapButtonContext } from '@cow/modules/swap/hooks/useSwapButtonContext'
-import { SwapFormProps } from '@cow/modules/swap/containers/NewSwapWidget/typings'
+import { SwapFormProps } from '@cow/modules/swap/containers/NewSwapWidget/types'
 import { ConfirmSwapModalSetupProps } from '@cow/modules/swap/containers/ConfirmSwapModalSetup'
 import { EthFlowProps } from '@cow/modules/swap/containers/EthFlow'
 import { NewSwapModals, NewSwapModalsProps } from '@cow/modules/swap/containers/NewSwapModals'
