+++ src/cow-react/modules/swap/pure/Row/RowFeeContent/index.tsx
@@ -3,7 +3,7 @@ import { MouseoverTooltipContent } from 'components/Tooltip'
 import { StyledInfo } from '@cow/pages/Swap/styleds'
 import { LightGreyText } from '@cow/modules/swap/pure/styled'
 import { StyledRowBetween, TextWrapper } from '@cow/modules/swap/pure/Row/styled'
-import { RowStyleProps, RowWithShowHelpersProps } from '@cow/modules/swap/pure/Row/typings'
+import { RowStyleProps, RowWithShowHelpersProps } from '@cow/modules/swap/pure/Row/types'
 
 export interface RowFeeContentProps extends RowWithShowHelpersProps {
   includeGasMessage: string
