+++ src/cow-react/modules/swap/pure/Row/RowSlippageContent/index.tsx
@@ -7,7 +7,7 @@ import { INPUT_OUTPUT_EXPLANATION } from 'constants/index'
 import { StyledInfo } from '@cow/pages/Swap/styleds'
 import { RowSlippageProps } from '@cow/modules/swap/containers/RowSlippage'
 import { StyledRowBetween, TextWrapper } from '@cow/modules/swap/pure/Row/styled'
-import { RowStyleProps } from '@cow/modules/swap/pure/Row/typings'
+import { RowStyleProps } from '@cow/modules/swap/pure/Row/types'
 
 const ClickableText = styled.button`
   background: none;
